chore: dedupe same function
